refactor: rename and move reducer file
